backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
langchain==0.1.11
langgraph==0.0.26
google-generativeai==0.3.2
pypdf==4.0.1
chromadb==0.4.24
sqlalchemy==2.0.27
python-jose==3.3.0
email-validator==2.1.0.post1
python-dotenv==1.0.1
pydantic==2.6.1
aiofiles==23.2.1
beautifulsoup4==4.12.3
jsonschema==4.21.1
httpx==0.26.0 